Package: telnet
New package: Telnet protocol client and server ported from OpenBSD
(post-3.0), with significant modifications.  The Telnet protocol
handling in telnetd is performed in a process running as a dedicated
pseudo-user and chrooted to /var/empty.  This uses the approach
introduced by Chris Evans in his NetKit telnetd patches, but the code
is different.  Please refer to change log for the package itself for
descriptions of the many modifications applied during the week this
package was in development.

CVS-ID: doc/CHANGES 1.52