Package: glibc SECURITY FIX    Severity: none to high, remote, active
Back-ported a glob(3) buffer overflow fix from the CVS.  The bug has
been discovered and an initial patch produced by Flavio Veloso of
Magnux.  While no Owl package is known to be affected by this glibc
bug, it is likely that it may result in a security hole with certain
third-party software such as FTP servers which support globbing and
make use of the glob(3) interface.  At the same time, asprintf(3) and
vasprintf(3) have been modified to behave on errors and match the
semantics of Todd Miller's implementation found on *BSD, -- thanks to
Dmitry V. Levin of ALT Linux for discovering and looking into these
issues.

CVS-ID: doc/CHANGES 1.56
CVS-ID: packages/glibc/glibc-2.1.3-cvs-20011129-glob.diff 1.1
CVS-ID: packages/glibc/glibc-2.1.3-owl-alt-asprintf-error-handling.diff 1.1
CVS-ID: packages/glibc/glibc.spec 1.23