Updated to 1.0.2.
Dropped Red Hat's autoconf/libtoolize patch.

Use the new Makefile-libbz2_so for building the shared library.

Package the bzip2 binary that is statically-linked against libbz2 for
better performance on register-starved architectures such as the x86.

Patched bzdiff/bzcmp (new with 1.0.2) to use mktemp(1) instead of
"tempfile" and to remove the temporary file in all cases.

Build with -Wall.

CVS-ID: packages/bzip2/bzip2-1.0.2-owl-Makefiles.diff 1.1
CVS-ID: packages/bzip2/bzip2-1.0.2-owl-tmp.diff 1.1
CVS-ID: packages/bzip2/bzip2.spec 1.8