Revert the previous change, it wasn't too smart for several reasons.
We'd require /usr/libexec/vitmp here but it makes sense to install this
package before an editor is installed or even without an editor at all.
What we would really like to require is that if a /bin/vi is installed,
there should also be a /usr/libexec/vitmp wrapper.  Unfortunately, this
can't easily be expressed within RPM's dependency handling.

CVS-ID: packages/quota/quota.spec 1.16
CVS-ID: packages/vixie-cron/vixie-cron.spec 1.15