Packages: vim, bash, quota, vixie-cron
Updated vim to 6.1 patchlevel 18, with various additional changes to
the package.  The package now includes vitmp(1), a wrapper around VIM
to be used for editing temporary files with in-place rewrites.  It is
now the default editor for crontab(1), edquota(8), the "fix command"
(fc) history editor in bash, and the bashbug script.

CVS-ID: doc/CHANGES 1.84