Package: glibc
ldd(1) will no longer try to invoke programs directly, even when it
seems like that would work.  The dynamic linker will be invoked as a
program instead.  This makes a difference primarily when the program
is SGID and is being ldd'ed by root.  If the program was executed
directly, glibc would detect its SGID status and drop LD_* variables,
resulting in the program being actually started rather than ldd'ed.
Thanks to Dmitry V. Levin of ALT Linux for suggesting this solution.
syslog(3) will now use ctime_r() instead of strftime_r() so that month
names will not depend on current locale settings.  The patch is
originally by Michael Tokarev.  The glibcbug script will now use
mktemp(1) in a fail-close way, let it use $TMPDIR, and will default to
vitmp(1) for the editor.

CVS-ID: doc/CHANGES 1.99