Package: openssh SECURITY FIX    Severity: none to high, remote, active
Updated to 3.4p1 with a lot of additional modifications to restore
most of the functionality lost or broken with the recent rushed update
to 3.3p1 and to be safer.  Please refer to change log for the package
for details.  OpenSSH 3.4p1 fixes the lack of bound checking resulting
in an integer and buffer overflow with the PAMAuthenticationViaKbdInt
code (and thus potentially allowing for a remote server compromise).
On Owl, PAMAuthenticationViaKbdInt has always defaulted to no and is
in fact not supported by our PAM configuration file for OpenSSH.  A
comment in /etc/ssh/sshd_config incorrectly seemed to imply that
PAMAuthenticationViaKbdInt defaulted to yes, which was never the case.
That comment has since been corrected.

Package: pam

pam_limits will now support stacking for account management (as well
as for session setup), be fail-close on configuration file reads, and
report the "too many logins" via PAM conversation rather than direct
printf(3).  The first change is needed for the new OpenSSH package.

CVS-ID: doc/CHANGES 1.108