Corrected the temporary file handling in perldoc (patch from ALT Linux) and Configure.
Use the versions of Perl-specific find-{provides,requires} included with
RPM, don't bring our own with this package.

Only generate *.ph files for packages which are a part of Owl, not other
packages which just happened to be installed on the build system, and make
the line producing STDH out of PKGS actually work (did they ever test this
at Red Hat? same bug in Rawhide, so it seems not).

Override myuname (to `uname -mrs` rather than `uname -a`) and myhostname
and don't package linux/compile.ph to not leak information specific to the
build system's last kernel compile.

Convention enforcement on some patch file names, dropped an obsolete patch.

CVS-ID: packages/perl/perl-5.6.0-alt-owl-perldoc-tmp.diff 1.1
CVS-ID: packages/perl/perl-5.6.0-owl-disable-suidperl.diff 1.1
CVS-ID: packages/perl/perl-5.6.0-owl-no-mail.diff 1.1
CVS-ID: packages/perl/perl-5.6.0-owl-tmp.diff 1.1
CVS-ID: packages/perl/perl-5.6.0-rh-install-man.diff 1.1
CVS-ID: packages/perl/perl-5.6.0-rh-no-db.diff 1.1
CVS-ID: packages/perl/perl.spec 1.9