Package: glibc SECURITY FIX    Severity: low to high, remote, passive to active
Patched two potential integer overflows (and thus buffer overflows) in
calloc(3) and Sun RPC xdr_array(3) code (the latter discovered by ISS
X-Force).  The calloc(3) integer overflow possibility is currently not
known to allow for an attack on a particular application, but has been
patched as a proactive measure.  The Sun RPC xdr_array(3) overflow may
allow for passive attacks on mount(8) by malicious or spoofed NFSv3
servers as well as for both passive and active attacks on RPC clients
or services that one might install on Owl.

CVS-ID: doc/CHANGES 1.118