Package: perl SECURITY FIX    Severity: none to high, remote, active
Back-ported bound checking fixes for File::Glob from Perl 5.8.0.
Thanks to Pavel Kankovsky for the report and to Michael Tokarev for
discussing other possible approaches to fixing this.  Without these
fixes, it was possible that certain otherwise correct Perl scripts
would expose the lack of bound checking in the Perl module code to be
exploited via user input to those scripts, which, depending on the
nature of such scripts, may be coming from a remote system.

CVS-ID: doc/CHANGES 1.123