Package: tar
Fixed the contains_dot_dot() bug introduced in 1.13.19 and discovered
by 3APA3A; thanks to Mark J Cox of Red Hat and Bencsath Boldizsar for
further analysis.  This is not marked a security fix because GNU tar
is still unsafe to use on untrusted archives for other reasons anyway.
Check signatures, then extract.

CVS-ID: doc/CHANGES 1.135
CVS-ID: packages/tar/tar-1.13.19-owl-dot-dot.diff 1.1
CVS-ID: packages/tar/tar.spec 1.11