Renamed CHANGES to CHANGES-1.0 for the current branch.  The intent is as follows:
CHANGES will remain in the repository because it was a shared file name
with the 0.1-stable branch.

CHANGES-1.0 will be a part of 1.0 release and its stable branch and it
may also stay in current and even in future releases for some time.

Separate CHANGES-* file names will be allocated for branches and to
accumulate changes leading up to future releases.

CVS-ID: doc/CHANGES-1.0 1.1