Our initial cvs package from Simon with only cosmetic changes.
This is still so broken that it can't even commit itself!  It SEGVs
in a vasprintf() call, to be investigated and fixed with subsequent
commits before I let this to the public.  The only reason I commit
this now is to make my significant changes to the patches separate
from the previous work.

CVS-ID: packages/cvs/cvs-1.11.5-alt-owl-tmp.diff 1.1
CVS-ID: packages/cvs/cvs-1.11.5-owl-no-checkin-update-prog.diff 1.1
CVS-ID: packages/cvs/cvs-1.11.5-owl-zlib.diff 1.1
CVS-ID: packages/cvs/cvs.spec 1.1