Re-worked much of the temporary file handling patch to make it actually do at least some of what it was supposed to; also patched the fail-open use in configure.
vasprintf() just can't be used like that, it returns an int and not the
string.  And this chunk wasn't required anyway.

"trap" can only be used once per signal.

Use "rm -f" on non-directories, not "rm -rf".

In cvsbug, use the $xs variable for exit code as the rest of the script
relies upon it.

This is still mostly untested (although I'm doing the commit with the new
package already).  Whoever reads this, -- please find the time and test
everything that's touched by the patch! ;-)

CVS-ID: packages/cvs/cvs-1.11.5-alt-owl-tmp.diff 1.2
CVS-ID: packages/cvs/cvs.spec 1.2