/var/lib/dhcp (the new fs root) must not be writable by anyone but root.
This commit should fix permissions on the directories and leases file
for dhcpd.  The placement and permissions of dhclient.leases are
probably still wrong (should we allocate another pseudo-user and
state subdirectory under /var/lib/dhcp once the client is able to
handle that? meanwhile, it has nothing to do with user dhcpd anyway).

CVS-ID: packages/dhcp/dhcp.spec 1.7