Change the convention for setting up temporary file removal traps in shell scripts.  Let's setup those traps only after files are created.  Rationale:
1. There's the risk of races internal to the shell implementation.  This is
asynchronous signal handlers accessing global variables!  The worst that
could theoretically happen (depending on shell implementation) is the "rm"
command getting invoked on a partial string, removing the wrong file or a
directory tree (e.g., "rm -rf /" instead of "rm -rf /tmp/something").  This
does not appear to be possible with current versions of bash in particular,
but that's hardly something we want to rely upon.

2. Even if we setup traps before files are created, like we used to do,
there's still a race where the file may be created but its name not yet
assigned to the variable.  So neither approach is perfect in that respect.

Also provide an example for creating temporary directories.

CVS-ID: doc/CONVENTIONS 1.19