Set the default (undecided) value of __libc_enable_secure with all bits set, except for the sign bit (someone else might use "-1" to mean "insecure", same as *BSD securelevel?)
Make sure bit 0 is always set the same as it used to be in unpatched glibc.

Be safe in case the kernel passes multiple instances of AT_*ID (this patch
might be a candidate for upstream inclusion).

(Sorry for the diff file having changed so heavily, -- Dmitry has been using
a different version of diff(1), not the one we have in Owl.)

CVS-ID: packages/glibc/glibc-2.3.5-owl-alt-sanitize-env.diff 1.4