Many minor changes, mostly based on a discussion on owl-devel, including:
Don't install a lilo.conf, instead provide a lilo.conf.sample and list
the real lilo.conf as a %ghost.

When checking whether LILO was installed, also check for lilo.conf
(which we used to do, and which makes sense only if we don't install a
lilo.conf right here).

Output all messages in %post to stdout, don't redirect some to stderr.
(Of course, programs that we run may output some of their messages to
stderr anyway.)  This is consistent with what we currently have in other
packages, such as owl-etc.

Reverted the %defattr change to be consistnet with most other packages.

CVS-ID: packages/lilo/lilo.spec 1.26