Switch to a heavily cut-down BSD license, up to the point of being "copyright only":
Redistribution and use in source and binary forms, with or without
modification, are permitted.

Don't include any of the numbered conditions, but do include the
warranty disclaimer.

That way, the license is both obviously less restrictive than what we
had so far (in fact, not restrictive at all), and it is also obviously
compatible with "full" (2-clause and "bigger") BSD licenses.

We have to keep two man pages - pam_passwdqc.8 and passwdqc.conf.5 -
under the 3-clause BSD license, so this change applies to the rest of
the package.

ldv@ ok

CVS-ID: packages/passwdqc/passwdqc/LICENSE 1.7
CVS-ID: packages/passwdqc/passwdqc/pwqcheck.1 1.9
CVS-ID: packages/passwdqc/passwdqc/pwqgen.1 1.6