Dropped two outdated comments:
Pre-compiled libext2fs.info is in fact included in 1.41.9 (so the comment about
it missing was outdated), but we no longer seem to have a need to regenerate it
(so dropped the comment instead of uncommenting the "rm" command).

"make -j2" worked fine during my debugging, so the comment saying "this package
cannot be built using parallel tasks" appeared to be wrong/obsolete.

Also dropped the --enable-htree-clear option to %configure, which resulted in:

configure: WARNING: unrecognized options: --enable-htree-clear

and in fact appeared to be unrecognized.

CVS-ID: packages/e2fsprogs/e2fsprogs.spec 1.53