Don't avoid loading of duplicate hashes when DB_WORDS is set.
Previously, when DB_WORDS is set, we would avoid loading of duplicate hashes
when the corresponding login names also matched.  Unfortunately, this was too
slow in some reasonable cases (e.g., with very large numbers of LM hashes),
because the duplicate hashes that we loaded (because of differing login names)
would result in excessive hash "collisions" in db->password_hash[].  (A way to
avoid that would be to use a custom hash function that would take not only
the password hash but also the login name into account.)

Another problem with the old approach was that a matching password hash and
login name did not imply that all of the GECOS field words and the home
directory pathname would match as well, so we could be dropping valuable
info.  (Of course, the check could be enhanced, and all of this info could
be rolled into a custom hash function to avoid excessive "collisions".)

CVS-ID: packages/john/john/src/loader.c 1.12