In single_process_pw(), stop processing as soon as we've cracked the password.
Previously, some additional candidate passwords could be generated and tried
against other password hashes with the same salt and, if any of those are
successfully cracked, then the same passwords would also be tried against all
other hashes.  That's the way "single crack" mode works, but it was meant that
way for reusing info from not-yet-cracked hashes only, where the extra tests
against other password hashes with the same salt would come for free.  When
the target password is already cracked, those extra tests are no longer free,
and thus we're avoiding them now.

This change might result in fewer passwords getting cracked in some cases, yet
the old behavior was unintended and inconsistent, so it is being corrected now.

CVS-ID: packages/john/john/src/single.c 1.17