When matching against the reversed new password, always pass the original non-reversed new password (possibly with a substring removed) into is_simple(), but remove or check the correct substring in is_based() considering that the matching is possibly being done against the reversed password.
Before this change, "user %2U" would be rejected (with the default settings)
as based on a dictionary word, even though "use %2U" would be accepted.  The
cause was that "U2% resu", when passed into is_simple() by is_based(), would
have its initial uppercase character not counted towards the number of
character classes, even though this was actually the last character (not
the first) in the original password.  With "use %2U", this call was avoided
because "use" was too short for the default match length of 4.  With this
change, both of these passwords are accepted with the default settings.

CVS-ID: packages/passwdqc/passwdqc/passwdqc_check.c 1.26