Drop faulty check for sparse files in tcb_is_suspect()
It was based on a wrong assumption that st_blksize indicates the size
of allocated blocks.  Also, the notion of sparse files does not apply
to filesystems with compression turned on.

The purpose of this check was to prevent some DoS attacks on root
invoking user management tools and on services doing authentication.
However, on a system with tcb shadow files, if group shadow access is
somehow compromised, those files may be directly written to by their
corresponding users (as well as made sparse).

Bug reported by Jim Darby <jim at jimbocorp.uklinux.net>.

CVS-ID: packages/tcb/tcb/libs/libtcb.c 1.9
CVS-ID: packages/tcb/tcb/ChangeLog 1.44