Dropped most LILO boot targets, leaving only two: "normal" and "rescue" (new names for what was previously known as "autodetect" and "custom").
Reduced the LILO menu timeout from 60 to 5 seconds.

Updated the boot menu message to reflect the changes to the menu items.

Increased the rootdelay= setting from 10 to 30 seconds now that its
meaning is slightly different with our magic root=/dev/cdrom setting (it's
the maximum number of 1-second retries, not an initial delay anymore).

Revised welcome-cdrom.sh in numerous ways: don't bother looking for kernel
sources (they won't be found in a separate "top level" directory anymore),
refer to the source code tree without calling it "userland" (now that it
contains the kernel as well), provide simpler and more correct instructions
for setting a Cyrillic font (now via a shell alias), use highlighting of
command names (instead of double-quotes) when outputting to a terminal.

CVS-ID: packages/owl-cdrom/owl-cdrom.spec 1.55