Corrected the way /etc/cron.d/vz is created such that /etc/cron.d's mtime is changed, which is needed for our crond to actually notice the file.
Added a cron job to update quota files (in case the system crashes).

Staggered the cron jobs (don't run more than one on a given minute).

When creating a new container, pass the "--numeric-owner -Sp" options to tar
extracting the template, in addition to the options that were used previously.

Set the new container's root directory permissions to 755 regardless of the
current umask and of permissions for "." or "/" that might be in the tarball.

In *set_ugid_quota.sh scripts, set the permissions on /etc/mtab to 644
regardless of the current umask.

CVS-ID: packages/vzctl/vzctl-3.0.23-owl-cron.diff 1.1
CVS-ID: packages/vzctl/vzctl-3.0.23-owl-mtab-mode.diff 1.1
CVS-ID: packages/vzctl/vzctl-3.0.23-owl-vps-create.diff 1.1
CVS-ID: packages/vzctl/vzctl.spec 1.6