Backported a fix for CVE-2010-4651.  The patch utility allowed ".." in pathnames, and it also allowed absolute pathnames, either of which could allow an attacker to create or modify arbitrary files outside of the intended directory tree using a specially-crafted patch file.  Our partial fix of 2011/02/02 did not address the absolute pathname case. References: https://bugzilla.redhat.com/show_bug.cgi?id=667529 http://www.openwall.com/lists/oss-security/2011/01/05/10 http://lists.gnu.org/archive/html/bug-patch/2010-12/msg00000.html http://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2010-4651
Backported a partial fix for CVE-2010-4651.  Since the fix turned out to
be incomplete, this change is not actually fixing CVE-2010-4651 yet.

Moved a few syntax highlighting related files from the vim-syntax to the
vim-enhanced subpackage to correct a packaging error where some files
in vim-enhanced were dependent upon files from vim-syntax, which is not
installed by default.

CVS-ID: doc/CHANGES-3.0-stable 1.1.2.2