Check for external mode's generate() early on, before we start logging.
Check for external mode's filter() when the mode is used along with another
cracking mode (and thus is only useful for filtering), and refuse to run if
filter() is missing.  Previously, in this case the external mode specification
would be silently ignored.  This problem was reported and a different patch
for it proposed by Aleksey Cherepanov - thanks!

Check if the external mode being used for filtering only also has generate();
print a warning if so.

CVS-ID: packages/john/john/src/external.c 1.7
CVS-ID: packages/john/john/src/external.h 1.2
CVS-ID: packages/john/john/src/options.c 1.23