Added prepare() format method. Pass a pointer to the format itself to init(), prepare(), valid(), split().
We're not making much use of this functionality in the core JtR tree, but the
equivalent functionality was introduced in -jumbo by JimF and it is being made
use of there.  Hence, this commit lets us reduce the number of changes to core
code in future -jumbo versions, and it lets us proceed with further format
interface revisions in core without having both of the trees deviate from a
common base.

CVS-ID: packages/john/john/src/AFS_fmt.c 1.5
CVS-ID: packages/john/john/src/BF_fmt.c 1.12
CVS-ID: packages/john/john/src/BSDI_fmt.c 1.18
CVS-ID: packages/john/john/src/DES_fmt.c 1.15
CVS-ID: packages/john/john/src/LM_fmt.c 1.19
CVS-ID: packages/john/john/src/MD5_fmt.c 1.10
CVS-ID: packages/john/john/src/bench.c 1.14
CVS-ID: packages/john/john/src/c3_fmt.c 1.18
CVS-ID: packages/john/john/src/dummy.c 1.6
CVS-ID: packages/john/john/src/formats.c 1.10
CVS-ID: packages/john/john/src/formats.h 1.9
CVS-ID: packages/john/john/src/loader.c 1.50
CVS-ID: packages/john/john/src/trip_fmt.c 1.10