The previous commit failed testing with:
	strace -e open postmap /etc/postfix/aliases

which was still opening the DB_CONFIG file in the current directory.

This commit revises the previous one to check for and cure the symptom of
opening DB_CONFIG in the current directory, rather than try and fail to
selectively eliminate implicit opening of DB_CONFIG only.  Now the above test
with postmap passes (the file is no longer opened).

It is yet unclear whether this would pass db4's own tests.  We don't and can't
run those on clean Owl as they appear to require tcl, which we lack.

CVS-ID: packages/db4/db-4.3.29-owl-DB_CONFIG.diff 1.2
CVS-ID: packages/db4/db4.spec 1.27